telemetry: fix loong64 build
